+++ src/FunFair.CodeAnalysis/Helpers/Rules.cs
@@ -36,5 +36,6 @@ namespace FunFair.CodeAnalysis.Helpers
         public const string RuleDontUseBuildInAddOrUpdateConcurrentDictionary = @"FFS0032";
         public const string RuleDontUseBuildInGetOrAddConcurrentDictionary = @"FFS0033";
         public const string RuleDontUseConfigurationBuilderAddJsonFileWithReload = @"FFS0034";
+        public const string RuleTestClassesShouldNotDefineMutableFields = @"FFS0035";
     }
 }